# LibraryManagementSystem_API/Business/Dtos/AuthDtos/LoginRequestDto.cs
namespace LibraryManagementSystem_API.Business.Dtos.AuthDtos
{
    public class LoginRequestDto
    {
        public string username { get; set; }
        public string password { get; set; }

    }
}