# LibraryManagementSystem_API/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LibraryManagementSystem_API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RefreshToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/mail/{username}": {
      "get": {
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Notification"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Text",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Date",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "Read",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/staff/manage-book/{id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/staff/manage-book/add-book": {
      "post": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Author",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Category",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Stock",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Position",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "ReleaseDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "Length",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Publisher",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Synopsis",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Ebook",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/staff/manage-book/edit-book/{id}": {
      "put": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Author",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Category",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Stock",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Position",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "ReleaseDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "Length",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Publisher",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Synopsis",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Ebook",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/staff/manage-book/delete-book/{id}": {
      "delete": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/staff/user-request/{id}": {
      "put": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Accept",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Reason",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/staff/upload-image": {
      "post": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/staff/get-image/{id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/staff/post-epub/{id}": {
      "post": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "LoginRequestDto": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "RefreshToken": {
      "type": "object",
      "properties": {
        "refreshtoken": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "RegisterRequestDto": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}